requirements.txt
asgiref==3.8.1
cffi==1.17.0
colorama==0.4.6
cryptography==43.0.0
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Unipath==1.1
uritemplate==4.1.1
